Develop local

+++ src/main/java/com/example/base_framework/service/PermissionService.java
@@ -4,6 +4,7 @@ import com.example.base_framework.entity.Permission;
 import com.example.base_framework.repository.PermissionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -13,10 +14,12 @@ public class PermissionService {
 
     private final PermissionRepository permissionRepository;
 
+    @Transactional(readOnly = true)
     public List<Permission> getAllPermissions() {
         return permissionRepository.findAll();
     }
 
+    @Transactional
     public Permission createPermission(Permission request) {
 
         if (permissionRepository.findByName(request.getName()).isPresent()) {
@@ -26,7 +29,11 @@ public class PermissionService {
         return permissionRepository.save(request);
     }
 
+    @Transactional
     public void deletePermission(Long id) {
+        if (!permissionRepository.existsById(id)) {
+            throw new RuntimeException("Permission not found");
+        }
         permissionRepository.deleteById(id);
     }
 }
